pyproject.toml
[project]
name = "peteollama-serverless"
version = "2.0.0"
description = "PeteOllama Serverless - Routes all requests through RunPod"
authors = [
    {name = "Pete", email = "pete@peteproperties.com"}
]
requires-python = ">=3.12"
dependencies = [
    # Core Framework (Required)
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    # HTTP Client (Required for RunPod API)
    "requests>=2.31.0",
    "httpx>=0.25.0",
    # Environment Management (Required)
    "python-dotenv>=1.0.0",
    "psutil>=7.0.0",
    "pendulum>=3.0.0",
    "beartype>=0.17.0",
    "rich>=14.1.0",
    "mcp>=1.13.0",
    "loguru>=0.7.2",
    "torch>=2.2.0",
    "transformers>=4.55.0",
    "sentence-transformers>=5.1.0",
    "faiss-cpu>=1.12.0",
    "numpy>=2.3.0",
    "scikit-learn>=1.7.0",
    "scipy>=1.16.0",
    "langchain-huggingface>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "aiohttp>=3.9.0",
    "python-multipart>=0.0.6",
    "websockets>=11.0.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.30",
    "pandas>=2.2.2",
    "ollama>=0.1.8",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.27",
    "langchain-ollama>=0.3.6",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "runpod>=1.7.13",
    "langgraph>=0.6.6",
    "langsmith>=0.4.14",
    "duckduckgo-search>=8.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]
